TST: Remove/fix ensembling tests
Most of the ensembling tests (e.g. test_ensemble_pdes, test_ensemble_SR3)
didn't actually test anything other than that the combination ran without
errors.  But since the EnsemblingOptimizer decouples a lot of code, it doesn't
interact with those options.

In some cases, it may be worth asking whether certain combinations of features,
e.g. library ensembling & SR3 constraints are actually possible.  But that
wasn't what was tested, so I would add that as a GH issue to investigate the
behavior.
